requirements.txt
altgraph==0.17.2
attrs==21.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
deprecation==2.1.0
dkpro-cassis==0.7.2
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.4.0
joblib==1.1.0
lxml==4.9.1
more-itertools==8.12.0
nltk==3.7
numpy==1.21.6
packaging==21.3
pandas==1.3.5
prompt-toolkit==3.0.30
git+https://github.com/JavierLuna/pycaprio.git@0.2.1#egg=pycaprio
pyfiglet==0.8.post1
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.7
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==23.2.0
regex==2022.6.2
requests==2.28.1
requests-toolbelt==0.9.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sortedcontainers==2.4.0
tenacity==5.1.5
termcolor==2.0.1
threadpoolctl==3.1.0
toposort==1.7
tqdm==4.64.1
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
xmi2conll==0.1.2
zipp==3.8.0
